requirements.txt
fastapi
uvicorn
jinja2
python-docx
python-multipart